+++ clemcore/clemgame/master.py
@@ -7,6 +7,7 @@ from typing import List, Dict, Tuple, Any, Union, final, Optional
 
 from clemcore import backends
 from clemcore.clemgame.environment import Action, GameEnvironment
+from clemcore.clemgame.errors import ParseError, GameError
 from clemcore.clemgame.registry import GameSpec
 from clemcore.clemgame.player import Player
 from clemcore.clemgame.recorder import NoopGameRecorder
@@ -14,66 +15,6 @@ from clemcore.clemgame.resources import GameResourceLocator
 from clemcore.utils.string_utils import to_pretty_json
 
 module_logger = logging.getLogger(__name__)
-
-
-class ResponseError(Exception):
-    """
-    General error class for problems with the player response.
-
-    Developers can introduce more specific error types by subclassing this error.
-    Alternatively, the 'key' attribute can be used to define more granular error types.
-    """
-
-    def __init__(self, reason: Optional[str] = None, response: Optional[str] = None, key: Optional[str] = None):
-        """
-        :param reason: (optional) a brief description of the cause
-        :param response: (optional) the player's response
-        :param key: (optional) a key word
-        """
-        super().__init__(reason)
-        self.reason = reason
-        self.response = response
-        self.key = key
-
-    def __str__(self):
-        return f"{self.__class__.__name__}: {self.reason}"
-
-
-class ProtocolError(ResponseError):
-    """Raised when a message does not follow the communication protocol expected by the game master."""
-    pass
-
-
-class ParseError(ProtocolError):
-    """
-    This error is supposed to be raised when player messages cannot be parsed or understood by the game master e.g.
-    because the response does not start with a specified prefix.
-
-    For example:
-        - taboo: clue giver messages should start with 'CLUE:'
-        - wordle: guesser messages should start with 'GUESS:'
-    """
-    pass
-
-
-class GameError(ResponseError):
-    """Raised when a verbal action of a player causes problems for advancing the game."""
-    pass
-
-
-class RuleViolationError(GameError):
-    """Raised when a verbal action of a player violates the specified game rules.
-
-    For example:
-        - taboo: mentioning the target word as the clue giver
-        - wordle: guessing words that are not exactly 5 letters long
-    """
-    pass
-
-
-class NotApplicableError(GameError):
-    """Raised when a verbal action of a player cannot be applied to advance the game state."""
-    pass
 
 
 class GameMaster(abc.ABC):
@@ -487,11 +428,11 @@ class EnvGameMaster(GameMaster):
     """Extended GameMaster, integrating a GameEnvironment as self-contained object for state management."""
 
     def __init__(
-        self,
-        game_spec: GameSpec,
-        experiment: dict,
-        player_models: List[backends.Model],
-        game_environment: GameEnvironment,
+            self,
+            game_spec: GameSpec,
+            experiment: dict,
+            player_models: List[backends.Model],
+            game_environment: GameEnvironment,
     ):
         """
         Args:
@@ -552,7 +493,7 @@ class EnvGameMaster(GameMaster):
                 read from the game's instances.json.
         """
         self._on_setup(**kwargs)
-        if self.players_by_names: # todo: why should this be empty here?
+        if self.players_by_names:  # todo: why should this be empty here?
             self.current_player = self.get_players()[self.current_player_idx]
 
     @abc.abstractmethod
